Senior-Project-Website_Add_Optimizer/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
python-jose[cryptography]==3.5.0
bcrypt==4.0.1
pydantic==2.11.7
python-dotenv==1.1.1
email-validator==2.2.0
alembic==1.16.5
psycopg2-binary==2.9.10
scikit-learn==1.7.1
joblib==1.5.1
